Saving to layout file works properly
