--- types/widget.go
@@ -1,13 +1,0 @@
-package types
-
-import (
-	"encoding/json"
-)
-
-// Widget interface
-type Widget interface {
-	json.Unmarshaler
-	Send(interface{})
-	Read()
-	Close() chan bool
-}
